Move `StateChange` to `txn_receipt.proto`
