+++ protos/txn_receipt.proto
@@ -19,7 +19,6 @@ option java_multiple_files = true;
 option java_package = "sawtooth.sdk.protobuf";
 option go_package = "txn_receipt_pb2";
 
-import "state_delta.proto";
 import "events.proto";
 
 
@@ -38,4 +37,17 @@ message TransactionReceipt {
   repeated Data data = 3;
 
   string transaction_id = 4;
+}
+
+// A state change is an entry in a given delta set. StateChange objects
+// have the type of SET, which is either an insert or update, or
+// DELETE.  Items marked as a DELETE will have no byte value.
+message StateChange {
+    enum Type {
+        SET = 0;
+        DELETE = 1;
+    }
+    string address = 1;
+    bytes value = 2;
+    Type type = 3;
 }
